build with non-dev sources
remote build script does not need to git stash/pop workaround when using
the non-development dependencies and auto-loader.

requires to not re-build the phar in the standard shell test run.

+++ test/shell/test-schema-validate.sh
@@ -23,8 +23,10 @@ case ${1-0} in
       ;;
   1 ) echo "# 1: build pipelines phar (validate with phar)"
       cd "${PROJECT_DIR}"
-      rm -f build/pipelines.phar
-      "${PHP_BINARY-php}" -d phar.readonly=0 -f lib/build/build.php | grep 'signature:'
+      if [ -z "${CI-}" ]; then # skip building the phar-file in CI
+        rm -f build/pipelines.phar
+        "${PHP_BINARY-php}" -d phar.readonly=0 -f lib/build/build.php | grep 'signature:'
+      fi
       build/pipelines.phar --version
       exit
       ;;
